fix(PL-2247): handle same version promotion
